configure.ac: for enable-all-crypto, enable_dilithium without regard for ENABLED_EXPERIMENTAL;
wolfssl/wolfcrypt/sha512.h: add a prototype for wc_Sha512HashBlock() (only implemented in wolfcrypt/src/port/riscv/riscv-64-sha512.c);

wolfcrypt/src/asn.c, wolfcrypt/src/wc_mlkem.c, wolfcrypt/src/wc_mlkem_poly.c: add casts (and fix a few type clashes) to suppress for -Wconversions.

+++ wolfcrypt/src/asn.c
@@ -18684,7 +18684,7 @@ int ConfirmSignature(SignatureCtx* sigCtx,
                         goto exit_cs;
                     }
                     if ((ret = wc_dilithium_set_level(sigCtx->key.dilithium,
-                            level)) < 0) {
+                            (byte)level)) < 0) {
                         goto exit_cs;
                     }
                     if ((ret = wc_Dilithium_PublicKeyDecode(key, &idx,
@@ -31914,7 +31914,7 @@ static int MakeSignature(CertSignCtx* certSignCtx, const byte* buf, word32 sz,
                 ret = wc_dilithium_sign_ctx_msg(NULL, 0, buf, sz, sig,
                     &outSz, dilithiumKey, rng);
                 if (ret == 0)
-                    ret = outSz;
+                    ret = (int)outSz;
             }
         }
     #endif /* HAVE_DILITHIUM && !WOLFSSL_DILITHIUM_NO_SIGN */
